Added Remove Book and Toggle Read Status Buttons
Added two new buttons. One that removed books from both the library array and the DOM when pressed. And another button that toggles the read status of the book in both the library array and in the DOM. Minor functionality of removing books needs troubleshooting to remap html button attribute for "data" to the correct index after removal of indices.